refactor(api): split MCP read tool modules
